requirements.txt
alembic==1.16.5
annotated-types==0.7.0
anyio==4.10.0
astroid==4.0.4
asyncpg==0.31.0
bcrypt==4.3.0
black==26.1.0
cachetools==6.2.4
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.4
click==8.2.1
colorama==0.4.6
cryptography==46.0.1
deprecation==2.1.0
dill==0.4.1
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
flake8==7.3.0
fsspec==2026.1.0
git-filter-repo==2.47.0
greenlet==3.2.4
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
isort==8.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
jwt==1.4.0
librt==0.8.1
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
mmh3==5.2.0
multidict==6.7.0
mypy==1.19.1
mypy_extensions==1.1.0
orjson==3.11.2
packaging==25.0
passlib==1.7.4
pathspec==1.0.4
platformdirs==4.9.2
postgrest==2.27.2
propcache==0.4.1
psycopg2==2.9.10
pyasn1==0.6.1
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.11.7
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyflakes==3.4.0
Pygments==2.19.2
pyiceberg==0.10.0
PyJWT==2.10.1
pylint==4.0.5
pyparsing==3.3.1
pyroaring==1.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytokens==0.4.1
PyYAML==6.0.2
realtime==2.27.2
requests==2.32.5
rich==14.1.0
rich-toolkit==0.15.0
rignore==0.6.4
rsa==4.9.1
sentry-sdk==2.35.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.43
starlette==0.47.2
storage3==2.27.2
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.27.2
supabase-auth==2.27.2
supabase-functions==2.27.2
tenacity==9.1.2
tomlkit==0.14.0
typer==0.16.1
typing-inspection==0.4.1
typing_extensions==4.14.1
ujson==5.11.0
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
yarl==1.22.0
